% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{bart_wrapper}
\alias{bart_wrapper}
\title{Wrapper for fitting dbarts}
\usage{
bart_wrapper(train, test, sigest = NA, sigdf = 3, sigquant = 0.9, k = 2,
  power = 2, base = 0.95, binaryOffset = 0, ntree = 200,
  ndpost = 1000, nskip = 100, printevery = 100, keepevery = 1,
  keeptrainfits = TRUE, usequants = FALSE, numcut = 100,
  printcutoffs = 0, nthread = 1, keepcall = TRUE, verbose = FALSE)
}
\arguments{
\item{train}{...}

\item{test}{...}

\item{test}{...}
}
\value{
A list

A list
}
\description{
Wrapper for fitting dbarts
}
\examples{
# TO DO: Add
#' @param train ...
# TO DO: Add
}
