% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtmle_functions.R
\name{cvauc_cvtmle}
\alias{cvauc_cvtmle}
\title{Compute CVTML estimates of cross-validated AUC}
\usage{
cvauc_cvtmle(Y, X, K, learner = "glm_wrapper", seed = 1234,
  parallel = FALSE, maxIter = 10, icTol = 1/length(Y), ...)
}
\arguments{
\item{Y}{The outcome}

\item{X}{The predictors}

\item{K}{The number of folds}

\item{learner}{The learner wrapper}

\item{seed}{A random seed to set}

\item{parallel}{Compute the predictors in parallel?}

\item{maxIter}{Maximum number of iterations for cvtmle}

\item{icTol}{Iterate until maxIter is reach or mean of cross-validated
efficient influence function is less than \code{icTol}}

\item{...}{other arguments, not currently used}
}
\value{
A list
}
\description{
TO DO: Add
}
\examples{
n <- 200
p <- 10
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
Y <- rbinom(n, 1, plogis(X[,1] + X[,10]))
fit <- cvauc_cvtmle(Y = Y, X = X, K = 5, learner = "glm_wrapper")
}
