% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtmle_functions.R
\name{.Dy}
\alias{.Dy}
\title{Compute a portion of the efficient influence function}
\usage{
.Dy(full_long_data, y)
}
\arguments{
\item{full_long_data}{A long form data set}

\item{y}{Which portion of the EIF to compute}
}
\value{
Vector of EIF
}
\description{
Compute a portion of the efficient influence function
}
