% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtmle_functions.R
\name{.estimatingFn}
\alias{.estimatingFn}
\title{An estimating function for cvAUC}
\usage{
.estimatingFn(auc = 0.5, prediction_list, gn)
}
\arguments{
\item{auc}{The value of auc to find root for}

\item{prediction_list}{Entry in prediction_list}

\item{gn}{Marginal probability of outcome}
}
\description{
An estimating function for cvAUC
}
