% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtmle_functions.R
\name{.estimatingFnNestedCV}
\alias{.estimatingFnNestedCV}
\title{An estimating function for cvAUC with cv initial estimates}
\usage{
.estimatingFnNestedCV(auc = 0.5, prediction_list, folds, gn, K)
}
\arguments{
\item{auc}{The value of auc to find root for}

\item{prediction_list}{Entry in prediction_list}

\item{gn}{Marginal probability of outcome}
}
\description{
An estimating function for cvAUC with cv initial estimates
}
