% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtmle_functions.R
\name{F_nBn_star}
\alias{F_nBn_star}
\title{Compute the (CV)TMLE cumulative dist at psi_x}
\usage{
F_nBn_star(psi_x, y, Psi_nBn_0, Y_Bn, epsilon = 0)
}
\arguments{
\item{psi_x}{Value to compute conditional (on Y=y) cdf of Psi}

\item{y}{Value of Y to condition on}

\item{Psi_nBn_0}{Values of Psi_nBn(X) from training sample}

\item{Y_Bn}{Values of Y from training sample}

\item{epsilon}{Vector of fluctuation parameter estimates}
}
\value{
Numeric value of CDF
}
\description{
Compute the (CV)TMLE cumulative dist at psi_x
}
