% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtmle_functions.R
\name{fluc_mod_optim_1}
\alias{fluc_mod_optim_1}
\title{Alternative fluctuation routine}
\usage{
fluc_mod_optim_1(epsilon, fld, tol = 0.001)
}
\arguments{
\item{epsilon}{Fluctuation parameter}

\item{fld}{full_long_data_list}
}
\description{
Alternative fluctuation routine
}
