% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{glm_wrapper}
\alias{glm_wrapper}
\title{Wrapper for fitting a main terms GLM}
\usage{
glm_wrapper(train, test)
}
\arguments{
\item{train}{...}

\item{test}{...}
}
\value{
A list
}
\description{
Wrapper for fitting a main terms GLM
}
\examples{
# TO DO: Add
}
