% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{polymars_wrapper}
\alias{polymars_wrapper}
\title{Wrapper for fitting polymars}
\usage{
polymars_wrapper(test, train)
}
\arguments{
\item{test}{...}

\item{train}{...}
}
\value{
A list
}
\description{
Wrapper for fitting polymars
}
\examples{
# TO DO: Add

}
