% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.cvauc}
\alias{print.cvauc}
\title{Print results of cvauc_cvtmle}
\usage{
\method{print}{cvauc}(x, ...)
}
\arguments{
\item{x}{An object of class "cvauc"}
}
\description{
Print results of cvauc_cvtmle
}
