% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{randomforest_wrapper}
\alias{randomforest_wrapper}
\title{Wrapper for fitting a main terms random forest}
\usage{
randomforest_wrapper(train, test, mtry = floor(sqrt(ncol(train$X))),
  ntree = 1000, nodesize = 1, maxnodes = NULL, importance = FALSE, ...)
}
\arguments{
\item{train}{...}

\item{test}{...}
}
\value{
A list
}
\description{
Wrapper for fitting a main terms random forest
}
\examples{
# TO DO: Add
}
