% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{ranger_wrapper}
\alias{ranger_wrapper}
\title{Wrapper for fitting a main terms random forest}
\usage{
ranger_wrapper(train, test, num.trees = 500,
  mtry = floor(sqrt(ncol(train$X))), write.forest = TRUE,
  probability = TRUE, min.node.size = 5, replace = TRUE,
  sample.fraction = ifelse(replace, 1, 0.632), num.threads = 1,
  verbose = TRUE, ...)
}
\arguments{
\item{train}{...}

\item{test}{...}

\item{mtry}{...}

\item{...}{...}

\item{ntree}{...}

\item{nodesize}{...}

\item{maxnodes}{...}

\item{importance}{...}
}
\value{
A list
}
\description{
Wrapper for fitting a main terms random forest
}
\examples{
# TO DO: Add
}
