% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{superlearner_wrapper}
\alias{superlearner_wrapper}
\title{Wrapper for fitting a main terms random forest}
\usage{
superlearner_wrapper(train, test, SL.library = c("SL.mean"), ...)
}
\arguments{
\item{train}{...}

\item{test}{...}

\item{SL.library}{super learner library}
}
\value{
A list
}
\description{
Wrapper for fitting a main terms random forest
}
\examples{
# TO DO: Add
}
