% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{svm_wrapper}
\alias{svm_wrapper}
\title{Wrapper for fitting svm}
\usage{
svm_wrapper(test, train, type.class = "nu-classification",
  kernel = "radial", nu = 0.5, degree = 3, cost = 1, coef0 = 0,
  ...)
}
\arguments{
\item{test}{...}

\item{train}{...}
}
\value{
A list
}
\description{
Wrapper for fitting svm
}
\examples{
# TO DO: Add

}
