% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{xgboost_wrapper}
\alias{xgboost_wrapper}
\title{Wrapper for fitting xgboost}
\usage{
xgboost_wrapper(test, train, ntrees = 500, max_depth = 4,
  shrinkage = 0.1, minobspernode = 2, params = list(), nthread = 1,
  verbose = 0, save_period = NULL)
}
\arguments{
\item{test}{...}

\item{train}{...}
}
\value{
A list
}
\description{
Wrapper for fitting xgboost
}
\examples{
# TO DO: Add

}
