% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{glmnet_wrapper}
\alias{glmnet_wrapper}
\title{Wrapper for fitting lasso}
\usage{
glmnet_wrapper(train, test, lambda.select = "ncoef", ncoef = max(c(1,
  trunc(min(c(sum(train$Y)/10, sum(1 - train$Y)/10))))))
}
\arguments{
\item{train}{...}

\item{test}{...}
}
\value{
A list
}
\description{
Wrapper for fitting lasso
}
\examples{
# TO DO: Add
}
